refactor: firebase delete 로직 분리
